Returns deviations from suggested export options

Given a secuTrialdata object, this function returns information on deviations from suggested export options.

check_export_options(dat)

Arguments

dat

secuTrialdata object

Details

While the package strives to allow loading of as many types of secuTrial data exports as possible, there are certain export options which are less likely to cause issues. If possible it is suggested to export data which adheres to a suggested option set. This function points out deviations from the suggested set of options which are:
is_zip == TRUE
refvals_separate == TRUE
add_id == TRUE
duplicate_meta == FALSE
encoding == "UTF-8"
form_status == TRUE
centre_info == TRUE
proj_setup == TRUE
dict_items$lang == "en"
hidden_fields == FALSE
structure == TRUE

Examples

path <- system.file("extdata", "sT_exports", "exp_opt",
                    "s_export_CSV-xls_CTU05_only_column_names.zip",
                    package = "secuTrialR")
sT_export <- read_secuTrial_raw(path)

secuTrialR:::check_export_options(sT_export)
#> The following export options deviate from the suggested specifications:
#> Add-ID is not part of the export.
#> Form status is not part of the export.
#> Centre information is not part of the export.
#> Project setup is not part of the export.
#> Structure information is not part of the export.
#> Short names was not selected.