Data dictionary for secuTrialdata objects

Data dictionary for secuTrialdata objects

dictionary_secuTrial(x)

Arguments

x

secuTrialdata object

Value

data frame with info on variables

Examples

export_location <- system.file("extdata", "sT_exports", "lnames",
                               "s_export_CSV-xls_CTU05_long_ref_miss_en_utf8.zip",
                               package = "secuTrialR")
# read all export data
sT_export <- read_secuTrial(data_dir = export_location)
#> Read export successfully.
#> The following export options deviate from the suggested specifications:
#> Data from hidden fields is part of the export.
#> Short names was not selected.
# get dictionary
dictionary_secuTrial(sT_export)
#>            formtablename               formname            ffcolname
#> 1     emnpctu05surgeries              Surgeries        surgery_organ
#> 2     emnpctu05surgeries              Surgeries         surgery_type
#> 3             mnpctu05ae         Adverse Events            ae_is_sae
#> 4             mnpctu05ae         Adverse Events          ae_end_time
#> 5             mnpctu05ae         Adverse Events       ae_description
#> 6             mnpctu05ae         Adverse Events                notes
#> 7             mnpctu05ae         Adverse Events           ae_outcome
#> 8             mnpctu05ae         Adverse Events        ae_onset_date
#> 9             mnpctu05ae         Adverse Events                 <NA>
#> 10            mnpctu05ae         Adverse Events          ae_end_date
#> 11            mnpctu05ae         Adverse Events         ae_intensity
#> 12            mnpctu05ae         Adverse Events      ae_outcome_spec
#> 13            mnpctu05ae         Adverse Events        ae_onset_time
#> 14            mnpctu05ae         Adverse Events           ae_ongoing
#> 15            mnpctu05ae         Adverse Events       ae_action_spec
#> 16            mnpctu05ae         Adverse Events     ae_drug_relation
#> 17            mnpctu05ae         Adverse Events         ae_frequency
#> 18            mnpctu05ae         Adverse Events            ae_action
#> 19       mnpctu05allmedi     Further medication           med_family
#> 20       mnpctu05allmedi     Further medication          med_product
#> 21       mnpctu05allmedi     Further medication         no_more_meds
#> 22      mnpctu05baseline         General screen           birth_year
#> 23      mnpctu05baseline         General screen                  age
#> 24      mnpctu05baseline         General screen               gender
#> 25      mnpctu05baseline         General screen             hiv_date
#> 26      mnpctu05baseline         General screen              aspirin
#> 27      mnpctu05baseline         General screen      hospitalisation
#> 28      mnpctu05baseline         General screen                  icu
#> 29      mnpctu05baseline         General screen          paracetamol
#> 30      mnpctu05baseline         General screen   treat_type_paracet
#> 31      mnpctu05baseline         General screen  paracetamol_unknown
#> 32      mnpctu05baseline         General screen    paracetamol_start
#> 33      mnpctu05baseline         General screen     paracetamol_stop
#> 34      mnpctu05baseline         General screen            ibuprofen
#> 35      mnpctu05baseline         General screen treat_type_ibuprofen
#> 36      mnpctu05baseline         General screen    ibuprofen_unknown
#> 37      mnpctu05baseline         General screen      ibuprofen_start
#> 38      mnpctu05baseline         General screen       ibuprofen_stop
#> 39      mnpctu05baseline         General screen         aspirin_stop
#> 40      mnpctu05baseline         General screen   treat_type_aspirin
#> 41      mnpctu05baseline         General screen      aspirin_unknown
#> 42      mnpctu05baseline         General screen        aspirin_start
#> 43      mnpctu05baseline         General screen              surgery
#> 44      mnpctu05baseline         General screen    other_comorb_type
#> 45      mnpctu05baseline         General screen     immunodeficiency
#> 46      mnpctu05baseline         General screen       hiv_date_uknwn
#> 47      mnpctu05baseline         General screen               weight
#> 48      mnpctu05baseline         General screen                  hiv
#> 49      mnpctu05baseline         General screen           visit_date
#> 50      mnpctu05baseline         General screen               height
#> 51      mnpctu05baseline         General screen    baseline_comments
#> 52      mnpctu05baseline         General screen         other_comorb
#> 53      mnpctu05baseline         General screen     no_clinical_data
#> 54      mnpctu05baseline         General screen          liver_cirrh
#> 55      mnpctu05baseline         General screen           painkiller
#> 56      mnpctu05baseline         General screen     liver_cirrh_type
#> 57       mnpctu05outcome              Follow-up            follow_up
#> 58       mnpctu05outcome              Follow-up           death_date
#> 59       mnpctu05outcome              Follow-up  outcome_description
#> 60           mnpctu05sae Serious Adverse Events       sae_onset_date
#> 61           mnpctu05sae Serious Adverse Events          sae_ongoing
#> 62           mnpctu05sae Serious Adverse Events         sae_end_date
#> 63           mnpctu05sae Serious Adverse Events         sae_end_time
#> 64           mnpctu05sae Serious Adverse Events       sae_onset_time
#> 65           mnpctu05sae Serious Adverse Events                 <NA>
#> 66           mnpctu05sae Serious Adverse Events    sae_drug_relation
#> 67           mnpctu05sae Serious Adverse Events        sae_frequency
#> 68           mnpctu05sae Serious Adverse Events           sae_action
#> 69           mnpctu05sae Serious Adverse Events      sae_action_spec
#> 70           mnpctu05sae Serious Adverse Events          sae_outcome
#> 71           mnpctu05sae Serious Adverse Events     sae_outcome_spec
#> 72           mnpctu05sae Serious Adverse Events      sae_description
#> 73           mnpctu05sae Serious Adverse Events                notes
#> 74           mnpctu05sae Serious Adverse Events        sae_intensity
#> 75 mnpctu05studyterminat      Study termination         term_planned
#> 76 mnpctu05studyterminat      Study termination        term_toxicity
#> 77 mnpctu05studyterminat      Study termination           term_death
#> 78     mnpctu05treatment              Treatment      rando_treatment
#>                             itemtype
#> 1                Popup (Label Group)
#> 2                Popup (Label Group)
#> 3             Horizontal Radiobutton
#> 4               Checked Time (hh:mm)
#> 5                      Textarea 9,60
#> 6                      Textarea 9,60
#> 7             Horizontal Radiobutton
#> 8          Checked Date (dd.mm.yyyy)
#> 9                       Layout Dummy
#> 10         Checked Date (dd.mm.yyyy)
#> 11            Horizontal Radiobutton
#> 12                     Textarea 9,60
#> 13              Checked Time (hh:mm)
#> 14            Horizontal Radiobutton
#> 15                     Textarea 9,60
#> 16            Horizontal Radiobutton
#> 17            Horizontal Radiobutton
#> 18            Horizontal Radiobutton
#> 19                      Catalogfield
#> 20                      Catalogfield
#> 21                          Checkbox
#> 22               Checked Date (yyyy)
#> 23         Years y (calculated only)
#> 24            Horizontal Radiobutton
#> 25   Checked Date (dd.mm.yyyy hh:mm)
#> 26                          Checkbox
#> 27            Horizontal Radiobutton
#> 28            Horizontal Radiobutton
#> 29                          Checkbox
#> 30            Horizontal Radiobutton
#> 31                          Checkbox
#> 32                 Date (dd.mm.yyyy)
#> 33                 Date (dd.mm.yyyy)
#> 34                          Checkbox
#> 35            Horizontal Radiobutton
#> 36                          Checkbox
#> 37                 Date (dd.mm.yyyy)
#> 38                 Date (dd.mm.yyyy)
#> 39                 Date (dd.mm.yyyy)
#> 40            Horizontal Radiobutton
#> 41                          Checkbox
#> 42                 Date (dd.mm.yyyy)
#> 43            Horizontal Radiobutton
#> 44                      Textfield 40
#> 45            Horizontal Radiobutton
#> 46                          Checkbox
#> 47                        Number 3,2
#> 48            Horizontal Radiobutton
#> 49         Checked Date (dd.mm.yyyy)
#> 50                        Number 3,1
#> 51                   Textarea 10,100
#> 52            Horizontal Radiobutton
#> 53                          Checkbox
#> 54            Horizontal Radiobutton
#> 55            Horizontal Radiobutton
#> 56               Popup (Label Group)
#> 57            Horizontal Radiobutton
#> 58                 Date (dd.mm.yyyy)
#> 59                   Textarea 10,100
#> 60         Checked Date (dd.mm.yyyy)
#> 61            Horizontal Radiobutton
#> 62         Checked Date (dd.mm.yyyy)
#> 63              Checked Time (hh:mm)
#> 64              Checked Time (hh:mm)
#> 65                      Layout Dummy
#> 66            Horizontal Radiobutton
#> 67            Horizontal Radiobutton
#> 68            Horizontal Radiobutton
#> 69                     Textarea 9,60
#> 70            Horizontal Radiobutton
#> 71                     Textarea 9,60
#> 72                     Textarea 9,60
#> 73                     Textarea 9,60
#> 74            Horizontal Radiobutton
#> 75                          Checkbox
#> 76                          Checkbox
#> 77                          Checkbox
#> 78 Configurable Randomization Button
#>                                                fflabel        unit
#> 1                                                Organ        <NA>
#> 2                                                 Type        <NA>
#> 3                                            Is an SAE        <NA>
#> 4                                           Timepoints        <NA>
#> 5                         Description of Adverse Event        <NA>
#> 6                                                Notes        <NA>
#> 7                                              Outcome        <NA>
#> 8                                          Onset of AE        <NA>
#> 9                                        AE properties        <NA>
#> 10                                           End of AE        <NA>
#> 11                                           Intensity        <NA>
#> 12                           If others, please specify        <NA>
#> 13                                          Timepoints        <NA>
#> 14                                             Ongoing        <NA>
#> 15 Specify therapy / concomitant medication / Comments        <NA>
#> 16                                   Relation to study        <NA>
#> 17                                           Frequency        <NA>
#> 18                                        Action taken        <NA>
#> 19                                              Family        <NA>
#> 20                                             Product        <NA>
#> 21                               No further medication        <NA>
#> 22                                          Birth year        <NA>
#> 23                                                 Age        <NA>
#> 24                                              Gender        <NA>
#> 25                                                 HIV        <NA>
#> 26                                             Aspirin     Aspirin
#> 27                                     Hospitalisation        <NA>
#> 28                                                 ICU        <NA>
#> 29                                         Paracetamol Paracetamol
#> 30                    prior treatment type Paracetamol        <NA>
#> 31                           prior Paracetamol unknown        <NA>
#> 32                             prior Paracetamol start        <NA>
#> 33                              prior Paracetamol stop        <NA>
#> 34                                           Ibuprofen   Ibuprofen
#> 35                      prior treatment type Ibuprofen        <NA>
#> 36                             prior Ibuprofen unknown        <NA>
#> 37                               prior Ibuprofen start        <NA>
#> 38                                prior Ibuprofen stop        <NA>
#> 39                                  prior Aspirin stop        <NA>
#> 40                        prior treatment type Aspirin        <NA>
#> 41                               prior Aspirin unknown        <NA>
#> 42                                 prior Aspirin start        <NA>
#> 43                          Surgery since last screen?        <NA>
#> 44                              Other comorbidity type        <NA>
#> 45                                    Immunodeficiency        <NA>
#> 46                                    HIV date unknown        <NA>
#> 47                                              Weight          kg
#> 48                                                 HIV        <NA>
#> 49                                       Date of visit        <NA>
#> 50                                              Height          cm
#> 51                                            Comments        <NA>
#> 52                                   Other comorbidity        <NA>
#> 53                                    No clinical data        <NA>
#> 54                                     Liver cirrhosis        <NA>
#> 55                         Treatment with painkillers?        <NA>
#> 56                                Liver cirrhosis type        <NA>
#> 57                                           Follow-up        <NA>
#> 58                                 Exact date of death        <NA>
#> 59                     Optional description of outcome        <NA>
#> 60                                        Onset of SAE        <NA>
#> 61                                             Ongoing        <NA>
#> 62                                          End of SAE        <NA>
#> 63                                          Timepoints        <NA>
#> 64                                          Timepoints        <NA>
#> 65                                      SAE properties        <NA>
#> 66                                   Relation to study        <NA>
#> 67                                           Frequency        <NA>
#> 68                                        Action taken        <NA>
#> 69 Specify therapy / concomitant medication / Comments        <NA>
#> 70                                             Outcome        <NA>
#> 71                           If others, please specify        <NA>
#> 72                Description of Serious Adverse Event        <NA>
#> 73                                               Notes        <NA>
#> 74                                           Intensity        <NA>
#> 75                                             Planned     planned
#> 76                                            Toxicity    toxicity
#> 77                                               Death       death
#> 78                                       Randomization        <NA>
#>            formfamily
#> 1            Subforms
#> 2            Subforms
#> 3      Adverse Events
#> 4      Adverse Events
#> 5      Adverse Events
#> 6      Adverse Events
#> 7      Adverse Events
#> 8      Adverse Events
#> 9      Adverse Events
#> 10     Adverse Events
#> 11     Adverse Events
#> 12     Adverse Events
#> 13     Adverse Events
#> 14     Adverse Events
#> 15     Adverse Events
#> 16     Adverse Events
#> 17     Adverse Events
#> 18     Adverse Events
#> 19 Further medication
#> 20 Further medication
#> 21 Further medication
#> 22             Visits
#> 23             Visits
#> 24             Visits
#> 25             Visits
#> 26             Visits
#> 27             Visits
#> 28             Visits
#> 29             Visits
#> 30             Visits
#> 31             Visits
#> 32             Visits
#> 33             Visits
#> 34             Visits
#> 35             Visits
#> 36             Visits
#> 37             Visits
#> 38             Visits
#> 39             Visits
#> 40             Visits
#> 41             Visits
#> 42             Visits
#> 43             Visits
#> 44             Visits
#> 45             Visits
#> 46             Visits
#> 47             Visits
#> 48             Visits
#> 49             Visits
#> 50             Visits
#> 51             Visits
#> 52             Visits
#> 53             Visits
#> 54             Visits
#> 55             Visits
#> 56             Visits
#> 57             Visits
#> 58             Visits
#> 59             Visits
#> 60     Adverse Events
#> 61     Adverse Events
#> 62     Adverse Events
#> 63     Adverse Events
#> 64     Adverse Events
#> 65     Adverse Events
#> 66     Adverse Events
#> 67     Adverse Events
#> 68     Adverse Events
#> 69     Adverse Events
#> 70     Adverse Events
#> 71     Adverse Events
#> 72     Adverse Events
#> 73     Adverse Events
#> 74     Adverse Events
#> 75  Study termination
#> 76  Study termination
#> 77  Study termination
#> 78          Treatment