A function to assess the status of forms

This function returns a data.frame informing on the status of saved forms per participant. There is a line for every combination of form type and participant id. The numbers are occurrence counts.

form_status_counts(object, ...)

Arguments

object

secuTrialdata object

...

further parameters

Value

data.frame informing on the status of saved forms per participant

Examples

# prepare path to example export
export_location <- system.file("extdata", "sT_exports", "snames",
                               "s_export_CSV-xls_CTU05_short_ref_miss_en_utf8.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- read_secuTrial(data_dir = export_location)
#> Read export successfully.
#> The following export options deviate from the suggested specifications:
#> Data from hidden fields is part of the export.

# get form status
form_status_counts(sT_export)
#>           pat_id       form_name completely_filled partly_filled empty
#> 1  RPACK-CBE-001        baseline                 3             0     0
#> 2  RPACK-CBE-002        baseline                 1             0     0
#> 3  RPACK-CBE-003        baseline                 1             0     0
#> 4  RPACK-CBE-004        baseline                 1             0     0
#> 5  RPACK-CBE-005        baseline                 1             2     0
#> 6  RPACK-INS-011        baseline                 1             0     0
#> 7  RPACK-INS-012        baseline                 3             0     0
#> 8  RPACK-INS-013        baseline                 1             0     0
#> 9  RPACK-INS-014        baseline                 1             0     0
#> 10 RPACK-INS-015        baseline                 1             0     0
#> 11 RPACK-USB-123        baseline                 0             1     0
#> 12 RPACK-CBE-001         outcome                 1             0     0
#> 13 RPACK-CBE-002         outcome                 3             0     0
#> 14 RPACK-CBE-003         outcome                 1             0     0
#> 15 RPACK-CBE-004         outcome                 1             0     0
#> 16 RPACK-CBE-005         outcome                 1             0     0
#> 17 RPACK-INS-011         outcome                 1             0     0
#> 18 RPACK-INS-012         outcome                 1             0     0
#> 19 RPACK-INS-013         outcome                 1             0     0
#> 20 RPACK-INS-014         outcome                 1             0     0
#> 21 RPACK-INS-015         outcome                 1             0     0
#> 22 RPACK-USB-123         outcome                 0             1     0
#> 23 RPACK-CBE-002       atoutcome                 1             0     0
#> 24 RPACK-CBE-005       atoutcome                 3             0     0
#> 25 RPACK-INS-012       atoutcome                 1             0     0
#> 26 RPACK-CBE-001       treatment                 1             0     0
#> 27 RPACK-CBE-002       treatment                 1             0     0
#> 28 RPACK-CBE-003       treatment                 1             0     0
#> 29 RPACK-CBE-004       treatment                 1             0     0
#> 30 RPACK-CBE-005       treatment                 1             0     0
#> 31 RPACK-INS-011       treatment                 1             0     0
#> 32 RPACK-INS-012       treatment                 1             0     0
#> 33 RPACK-INS-013       treatment                 1             0     0
#> 34 RPACK-INS-014       treatment                 1             0     0
#> 35 RPACK-INS-015       treatment                 1             0     0
#> 36 RPACK-USB-123       treatment                 1             0     0
#> 37 RPACK-CBE-001         allmedi                 3             0     0
#> 38 RPACK-CBE-002         allmedi                 1             0     0
#> 39 RPACK-CBE-003         allmedi                 1             0     0
#> 40 RPACK-CBE-004         allmedi                 1             0     0
#> 41 RPACK-CBE-005         allmedi                 3             0     0
#> 42 RPACK-INS-011         allmedi                 1             0     0
#> 43 RPACK-INS-012         allmedi                 3             0     0
#> 44 RPACK-INS-013         allmedi                 1             0     0
#> 45 RPACK-INS-014         allmedi                 1             0     0
#> 46 RPACK-INS-015         allmedi                 1             0     0
#> 47 RPACK-USB-123         allmedi                 0             1     0
#> 48 RPACK-CBE-005       atallmedi                 2             0     0
#> 49 RPACK-CBE-001   studyterminat                 1             0     0
#> 50 RPACK-CBE-002   studyterminat                 1             0     0
#> 51 RPACK-CBE-003   studyterminat                 1             0     0
#> 52 RPACK-CBE-004   studyterminat                 1             0     0
#> 53 RPACK-CBE-005   studyterminat                 1             0     0
#> 54 RPACK-INS-011   studyterminat                 1             0     0
#> 55 RPACK-INS-012   studyterminat                 1             0     0
#> 56 RPACK-INS-013   studyterminat                 1             0     0
#> 57 RPACK-INS-014   studyterminat                 1             0     0
#> 58 RPACK-INS-015   studyterminat                 1             0     0
#> 59 RPACK-CBE-001 atstudyterminat                 1             0     0
#> 60 RPACK-INS-015              ae                 1             0     0
#> 61 RPACK-CBE-001             sae                 1             0     0
#> 62 RPACK-INS-014             sae                 1             0     0
#>    with_warnings with_errors
#> 1              0           0
#> 2              0           0
#> 3              0           0
#> 4              0           0
#> 5              0           0
#> 6              0           0
#> 7              0           0
#> 8              0           0
#> 9              0           0
#> 10             0           0
#> 11             0           0
#> 12             0           0
#> 13             0           0
#> 14             0           0
#> 15             0           0
#> 16             0           0
#> 17             0           0
#> 18             0           0
#> 19             0           0
#> 20             0           0
#> 21             0           0
#> 22             0           0
#> 23             0           0
#> 24             0           0
#> 25             0           0
#> 26             0           0
#> 27             0           0
#> 28             0           0
#> 29             0           0
#> 30             0           0
#> 31             0           0
#> 32             0           0
#> 33             0           0
#> 34             0           0
#> 35             0           0
#> 36             0           0
#> 37             0           0
#> 38             0           0
#> 39             0           0
#> 40             0           0
#> 41             0           0
#> 42             0           0
#> 43             0           0
#> 44             0           0
#> 45             0           0
#> 46             0           0
#> 47             0           0
#> 48             0           0
#> 49             0           0
#> 50             0           0
#> 51             0           0
#> 52             0           0
#> 53             0           0
#> 54             0           0
#> 55             0           0
#> 56             0           0
#> 57             0           0
#> 58             0           0
#> 59             0           0
#> 60             0           0
#> 61             0           0
#> 62             0           0