A function to show summary statistics for form statuses

This function warps form_status_counts and returns a data.frame summarizing the statuses for each form. Only saved forms are considered for the statistic.

form_status_summary(object, ...)

Arguments

object
secuTrialdata object

...
further parameters

Value

data.frame summarizing the statuses for each form

Examples

# prepare path to example export
export_location <- system.file("extdata","sT_exports", "snames",
                               "s_export_CSV-xls_CTU05_short_ref_miss_en_utf8.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- read_secuTrial(data_dir = export_location)

#> Read export successfully.

#> The following export options deviate from the suggested specifications:
#> Data from hidden fields is part of the export.

# get form status
form_status_summary(sT_export)

#>          form_name partly_filled completely_filled empty with_warnings
#> 1               ae             0                 1     0             0
#> 2          allmedi             1                16     0             0
#> 3        atallmedi             0                 2     0             0
#> 4        atoutcome             0                 5     0             0
#> 5  atstudyterminat             0                 1     0             0
#> 6         baseline             3                14     0             0
#> 7          outcome             1                12     0             0
#> 8              sae             0                 2     0             0
#> 9    studyterminat             0                10     0             0
#> 10       treatment             0                11     0             0
#>    with_errors partly_filled.percent completely_filled.percent empty.percent
#> 1            0            0.00000000                 1.0000000             0
#> 2            0            0.05882353                 0.9411765             0
#> 3            0            0.00000000                 1.0000000             0
#> 4            0            0.00000000                 1.0000000             0
#> 5            0            0.00000000                 1.0000000             0
#> 6            0            0.17647059                 0.8235294             0
#> 7            0            0.07692308                 0.9230769             0
#> 8            0            0.00000000                 1.0000000             0
#> 9            0            0.00000000                 1.0000000             0
#> 10           0            0.00000000                 1.0000000             0
#>    with_warnings.percent with_errors.percent form_count
#> 1                      0                   0          1
#> 2                      0                   0         17
#> 3                      0                   0          2
#> 4                      0                   0          5
#> 5                      0                   0          1
#> 6                      0                   0         17
#> 7                      0                   0         13
#> 8                      0                   0          2
#> 9                      0                   0         10
#> 10                     0                   0         11