This function loads a secuTrial export

This function will always load the full set of meta and data tables. The export options are also loaded and written into export_options.

read_secuTrial_raw(data_dir, ...)

# S3 method for secuTrialdata
print(x, ...)

Arguments

data_dir

string The data_dir specifies the path to the secuTrial data export.

...

further parameters

x

secuTrialdata object as returned by read_secuTrial_raw

Value

secuTrialdata object containing a list of export options and data.frames with all the data loaded from the secuTrial export. The list will contain at least the metadata data.frames and export_options list. data.frame with a row for each table in the export. For each table it contains the name, number of rows and columns, an indicator for whether the table is a metadata table and the files original name.

Examples

# prepare path to example export
export_location <- system.file("extdata", "sT_exports", "BMD",
                               "s_export_CSV-xls_BMD_short_en_utf8.zip",
                               package = "secuTrialR")
# read all export data
sT_export <- read_secuTrial_raw(data_dir = export_location)

# Print method
print(sT_export)
#> secuTrial data imported from:
#> /Users/runner/work/_temp/Library/secuTrialR/extdata/sT_exports/BMD/s_export_CSV-
#> xls_BMD_short_en_utf8.zip 
#>  table nrow ncol  meta original_name
#>     vp    1   10  TRUE        vp.xls
#>   vpfs    1    2  TRUE      vpfs.xls
#>     fs    1    7  TRUE        fs.xls
#>     qs    1    7  TRUE        qs.xls
#>     is    3    8  TRUE        is.xls
#>    ctr    1    3  TRUE       ctr.xls
#>     cn  113   13  TRUE        cn.xls
#>   atcn    0    6  TRUE      atcn.xls
#>  atcvp    0   11  TRUE     atcvp.xls
#>    qac    0   10  TRUE       qac.xls
#>    cts    0    8  TRUE       cts.xls
#>    bmd  504   27 FALSE       bmd.xls
#>  atbmd    0   28 FALSE     atbmd.xls
# or
sT_export
#> secuTrial data imported from:
#> /Users/runner/work/_temp/Library/secuTrialR/extdata/sT_exports/BMD/s_export_CSV-
#> xls_BMD_short_en_utf8.zip 
#>  table nrow ncol  meta original_name
#>     vp    1   10  TRUE        vp.xls
#>   vpfs    1    2  TRUE      vpfs.xls
#>     fs    1    7  TRUE        fs.xls
#>     qs    1    7  TRUE        qs.xls
#>     is    3    8  TRUE        is.xls
#>    ctr    1    3  TRUE       ctr.xls
#>     cn  113   13  TRUE        cn.xls
#>   atcn    0    6  TRUE      atcn.xls
#>  atcvp    0   11  TRUE     atcvp.xls
#>    qac    0   10  TRUE       qac.xls
#>    cts    0    8  TRUE       cts.xls
#>    bmd  504   27 FALSE       bmd.xls
#>  atbmd    0   28 FALSE     atbmd.xls