Write secuTrial exports to other formats

Convert the export prepared in R and export it to SPSS (sav), Stata (dta) or SAS (sas, xpt version 8) using the haven package.

write_secuTrial(object, ...)

# S3 method for secuTrialdata
write_secuTrial(object, format = "dta", metadata = FALSE, ...)

# S3 method for data.frame
write_secuTrial(df, filename, path = "", format = "dta", ...)

Arguments

object

secuTrialdata object

...

further parameters

format

format in which to save the export (one of "dta", "sas", "sav", "xpt")

metadata

if TRUE then metadate files will also be written

df

a data.frame

filename

file name

path

directory where the files should be saved

Value

a list of filenames

Details

Due to variable naming limitations in other packages, date variables are appended with _d (rather than _date), datetime/POSIX variables are appended with _dt (rather than _datetime) and factors with _f (rather than _factor). Further variable names may be altered in the conversion process. For details please refer to the haven documentation.

Examples

# prepare path to example export
export_location <- system.file("extdata", "sT_exports", "lnames",
                               "s_export_CSV-xls_CTU05_long_ref_miss_en_utf8.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- read_secuTrial(data_dir = export_location)
#> Read export successfully.
#> The following export options deviate from the suggested specifications:
#> Data from hidden fields is part of the export.
#> Short names was not selected.
tdir <- tempdir()
write_secuTrial(sT_export, format = "dta", path = tdir)
#> [[1]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/atmiv.dta"
#> 
#> [[2]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/ctu05baseline.dta"
#> 
#> [[3]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/atmnpctu05baseline.dta"
#> 
#> [[4]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/ctu05outcome.dta"
#> 
#> [[5]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/atmnpctu05outcome.dta"
#> 
#> [[6]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/ctu05treatment.dta"
#> 
#> [[7]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/atmnpctu05treatment.dta"
#> 
#> [[8]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/ctu05allmedi.dta"
#> 
#> [[9]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/atmnpctu05allmedi.dta"
#> 
#> [[10]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/ctu05studyterminat.dta"
#> 
#> [[11]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/atmnpctu05studyterminat.dta"
#> 
#> [[12]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/ctu05ae.dta"
#> 
#> [[13]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/atmnpctu05ae.dta"
#> 
#> [[14]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/ctu05sae.dta"
#> 
#> [[15]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/atmnpctu05sae.dta"
#> 
#> [[16]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/emnpctu05surgeries.dta"
#> 
#> [[17]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/atemnpctu05surgeries.dta"
#> 
#> [[18]]
#> [1] "Saved to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp1grw9B/atadverseevents.dta"
#> 
list.files(tdir)
#>  [1] "atadverseevents.dta"         "atemnpctu05surgeries.dta"   
#>  [3] "atmiv.dta"                   "atmnpctu05ae.dta"           
#>  [5] "atmnpctu05allmedi.dta"       "atmnpctu05baseline.dta"     
#>  [7] "atmnpctu05outcome.dta"       "atmnpctu05sae.dta"          
#>  [9] "atmnpctu05studyterminat.dta" "atmnpctu05treatment.dta"    
#> [11] "ctu05ae.dta"                 "ctu05allmedi.dta"           
#> [13] "ctu05baseline.dta"           "ctu05outcome.dta"           
#> [15] "ctu05sae.dta"                "ctu05studyterminat.dta"     
#> [17] "ctu05treatment.dta"          "downlit"                    
#> [19] "emnpctu05surgeries.dta"      "file3cac153ee4da"           
#> [21] "file3cac2aec506b"            "file3cac311640c4"           
#> [23] "file3cac47c79bbe"            "file3cac5422af1b"           
#> [25] "file3cac61c753a2"