as.data.frame method for secuTrialdata objects Make the data from the exports more easily accessible by placing them in another environment (e.g. place them in the global environment (.GlobalEnv) and you can reference them without referring to the secuTrialdata object anymore. Ie. they become regular data.frames).

as.data.frame method for secuTrialdata objects Make the data from the exports more easily accessible by placing them in another environment (e.g. place them in the global environment (.GlobalEnv) and you can reference them without referring to the secuTrialdata object anymore. Ie. they become regular data.frames).

# S3 method for secuTrialdata
as.data.frame(
  x,
  ...,
  envir,
  data.frames = NULL,
  meta = FALSE,
  regex = NULL,
  rep = ""
)

Arguments

x
secuTrialdata object

...
further parameters

envir
environment in which to put the data (e.g. .GlobalEnv)

data.frames
character vector of data.frame names to turn into data.frames

meta
logical should metadata be returned

regex
regex syntax to remove from names

rep
replacement for regex

Value

each data.frame in the secuTrialdata object is saved to it's own data.frame in the designated environment

Details

envir must be specifically defined. For simplicity, .GlobalEnv would probably be the easiest (assigning it to another environment would still entail referring to that environment).

Examples

# prepare path to example export
export_location <- system.file("extdata", "sT_exports", "lnames",
                               "s_export_CSV-xls_CTU05_long_ref_miss_en_utf8.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- read_secuTrial_raw(data_dir = export_location)
# add files to a new environment called env1
env1 <- new.env()
as.data.frame(sT_export, envir = env1)
# add files to a new environment called env2, removing the project name from
# the file names
env2 <- new.env()
as.data.frame(sT_export, regex = "ctu05", envir = env2)