Write secuTrial exports to other formats

Convert the export prepared in R and export it to SPSS (sav), Stata (dta) or SAS (sas, xpt version 8) using the haven package.

write_secuTrial(object, ...)

Arguments

object
secuTrialdata object

path
directory where the files should be saved

format
format in which to save the export (one of "dta", "sas", "sav", "xpt")

Value

a list of filenames

Details

Due to variable naming limitations in other packages, date variables are appended with _d (rather than _date), datetime/POSIX variables are appended with _dt (rather than _datetime) and factors with _f (rather than _factor). Further variable names may be altered in the conversion process. For details please refer to the haven documentation.

Examples

# prepare path to example export
export_location <- system.file("extdata", "sT_exports", "longnames",
                               "s_export_CSV-xls_CTU05_long_ref_miss_en_utf8.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- read_secuTrial(data_dir = export_location)

#> Read export successfully.

tdir <- tempdir()
write_secuTrial(sT_export, format = "dta", path = tdir)

#> [[1]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/atmiv.dta"
#> 
#> [[2]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/ctu05baseline.dta"
#> 
#> [[3]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/atmnpctu05baseline.dta"
#> 
#> [[4]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/ctu05outcome.dta"
#> 
#> [[5]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/atmnpctu05outcome.dta"
#> 
#> [[6]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/ctu05treatment.dta"
#> 
#> [[7]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/atmnpctu05treatment.dta"
#> 
#> [[8]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/ctu05allmedi.dta"
#> 
#> [[9]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/atmnpctu05allmedi.dta"
#> 
#> [[10]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/ctu05studyterminat.dta"
#> 
#> [[11]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/atmnpctu05studyterminat.dta"
#> 
#> [[12]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/ctu05ae.dta"
#> 
#> [[13]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/atmnpctu05ae.dta"
#> 
#> [[14]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/ctu05sae.dta"
#> 
#> [[15]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/atmnpctu05sae.dta"
#> 
#> [[16]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/emnpctu05surgeries.dta"
#> 
#> [[17]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/atemnpctu05surgeries.dta"
#> 
#> [[18]]
#> [1] "Saved to /tmp/Rtmp4qMTmZ/atadverseevents.dta"
#> 

list.files(tdir)

#>  [1] "atadverseevents.dta"             "atemnpctu05surgeries.dta"       
#>  [3] "atmiv.dta"                       "atmnpctu05ae.dta"               
#>  [5] "atmnpctu05allmedi.dta"           "atmnpctu05baseline.dta"         
#>  [7] "atmnpctu05outcome.dta"           "atmnpctu05sae.dta"              
#>  [9] "atmnpctu05studyterminat.dta"     "atmnpctu05treatment.dta"        
#> [11] "ctu05ae.dta"                     "ctu05allmedi.dta"               
#> [13] "ctu05baseline.dta"               "ctu05outcome.dta"               
#> [15] "ctu05sae.dta"                    "ctu05studyterminat.dta"         
#> [17] "ctu05treatment.dta"              "emnpctu05surgeries.dta"         
#> [19] "file5dd72671ad51.so"             "libloc_169_8f48fc7d1db1bb3e.rds"
#> [21] "libloc_174_173b0d0583319ed2.rds" "libloc_190_a664afc79c9c884b.rds"
#> [23] "libloc_198_906ea0066f46c4fa.rds"